router adjustment done
