+++ package.json
@@ -6,6 +6,8 @@
     "axios": "^0.19.2",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "1.0.13"
   },
   "scripts": {
