+++ src/containers/Blog/Blog.css
@@ -1,7 +1,23 @@
-.Posts {
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: center;
-    width: 80%;
-    margin: auto;
+.Blog ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  width: 100%;
+  text-align: center;
+  background: aquamarine;
+}
+
+.Blog li {
+  display: inline-block;
+  margin: 20px;
+}
+
+.Blog a {
+  text-decoration: none;
+  color: black;
+}
+
+.Blog a:hover,
+.Blog a:active {
+  color: #fa923f;
 }
