+++ src/containers/Blog/FullPost/FullPost.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react';
 import axios from 'axios';
-
 import './FullPost.css';
 
 class FullPost extends Component {
